refactor: oracles cleanup
