+++ noir-projects/aztec-nr/aztec/src/oracle/mod.nr
@@ -12,7 +12,6 @@ pub mod get_public_data_witness;
 pub mod get_membership_witness;
 pub mod keys;
 pub mod key_validation_request;
-pub mod get_sibling_path;
 pub mod random;
 pub mod enqueue_public_function_call;
 pub mod block_header;
