+++ yarn-project/circuit-types/src/interfaces/nullifier_membership_witness.ts
